requirements.txt
groq
google-play-scraper
python-dotenv
pytest
google-generativeai
markdown
fastapi
uvicorn[standard]
